requirements.txt
flask>=3.0.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
dnspython>=2.4.0
python-dotenv>=1.0.0
resend>=0.7.0
twilio>=8.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
